Softomation/HighwaySolutions/PySwbApp/requirements.txt
altgraph==0.17.4
appdirs==1.4.4
argcomplete==3.3.0
auto-py-to-exe==2.44.1
bcrypt==4.1.2
bidict==0.23.1
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
comtypes==1.4.1
cryptography==42.0.5
Cython==3.0.10
decorator==4.4.2
Deprecated==1.2.14
dnspython==2.6.1
Eel==0.16.0
eventlet==0.36.1
fabric==3.2.2
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
gmssl==3.2.2
GPUtil==1.4.0
greenlet==3.0.3
h11==0.14.0
hidapi==0.14.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
imutils==0.5.4
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.3.0
MarkupSafe==2.1.5
moviepy==1.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
netifaces==0.11.0
netifaces-plus==0.12.1
numpy==1.26.4
opencv-python==4.5.5.64
openvino==2023.3.0
openvino-telemetry==2024.1.0
packaging==24.0
paho-mqtt==1.6.1
pandas==2.2.2
paramiko==3.4.0
pefile==2023.2.7
pillow==10.3.0
ping3==4.0.8
proglog==0.1.10
psutil==2.1.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydub==0.25.1
pygame==2.5.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.8.0
pymodbus==3.7.0
PyNaCl==1.5.0
pynput==1.7.7
pyodbc==5.2.0
pyparsing==3.1.2
pypiwin32==223
pypng==0.20220715.0
Pypubsub==4.0.3
pyserial==3.5
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-escpos==3.1
python-socketio==5.11.2
pyttsx3==2.90
pytz==2024.1
pyudev==0.24.3
pyusb==1.2.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
QtAwesome==1.3.1
QtPy==2.4.1
reportlab==4.2.2
requests==2.31.0
#RFIDReaderAPI @ file:///D:/GitProjects/HighwaySolutions/Softomation/HighwaySolutions/WindowApplication/PyLane/libs/RFIDReaderAPI-1.1-py3-none-any.whl#sha256=b407f38187825e3d93e5da5d21cf40b92a7222e33376de6d84f16033887c68e9
scapy==2.5.0
schedule==1.2.1
scipy==1.14.0
shiboken6==6.6.1
simple-websocket==1.0.0
six==1.16.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vidgear==0.3.2
websockets==12.0
Werkzeug==3.0.2
whichcraft==0.6.1
wrapt==1.16.0
wsproto==1.2.0
xmlsec==1.3.14
zope.event==5.0
zope.interface==6.4.post2
